+++ Mlem/App/Views/Root/Login/LoginVersionWarningView.swift
@@ -15,24 +15,11 @@ struct LoginVersionWarningView: View {
 
     var body: some View {
         ScrollView {
-            VStack(alignment: .leading) {
-                Image(systemName: "xmark")
-                    .resizable()
-                    .symbolVariant(.square.fill)
-                    .aspectRatio(contentMode: .fit)
-                    .frame(height: 70)
-                    .foregroundStyle(.themedColorfulAccent(5))
-                    .padding(.top, 16)
-
-                Text("\(content.host) is unsupported")
-                    .font(.title)
-                    .fontWeight(.semibold)
-                    .padding(.bottom, 5)
-                if let software = content.software {
-                    Text(bodyText(software: software))
-                }
-            }
-            .padding(.horizontal, 16)
+            UnsupportedVersionDescriptionView(
+                host: content.host,
+                software: content.software,
+                offerContinuation: content.instance != nil
+            )
         }
         .frame(maxWidth: .infinity, maxHeight: .infinity)
         .background(.themedGroupedBackground)
@@ -45,26 +32,5 @@ struct LoginVersionWarningView: View {
                 }
             }
         }
-    }
-
-    func bodyText(software: SiteSoftware) -> String {
-        var result = String(localized: """
-                                       \(content.host) is running \(software.label), \
-                                       and Mlem requires \(minimumSoftware(type: software.type).label) or later.
-                                       """)
-        result += "\n\n"
-        result += .init(localized: "Consider choosing another instance, or asking your server administrators to upgrade.")
-        if content.instance != nil {
-            result += "\n\n"
-            result += .init(localized: "You can choose to continue anyway, but some features may not work on this version.")
-        }
-        return result
-    }
-
-    func minimumSoftware(type: SiteSoftwareType) -> SiteSoftware {
-        .init(
-            type: type,
-            version: type.minimumSupportedVersion
-        )
     }
 }
