+++ Mlem/App/Views/Shared/UnsupportedVersionWarningView.swift
@@ -5,15 +5,22 @@
 //  Created by Sjmarf on 2026-06-06.
 //
 
+import MlemMiddleware
 import SwiftUI
 
 struct UnsupportedVersionWarningView: View {
     let account: any Account
 
     var body: some View {
-        VStack {
-            Text("Unsupported")
+        ScrollView {
+            UnsupportedVersionDescriptionView(
+                host: account.host,
+                software: account.siteSoftware,
+                offerContinuation: true
+            )
         }
+        .frame(maxWidth: .infinity, maxHeight: .infinity)
+        .background(.themedGroupedBackground)
         .onDisappear {
             account.ignoreVersionWarning(true)
         }
